refactor(heatmap): extract HeatmapFileManager + HeatmapExporter (#197 Phase 2-3)
P1.4 Phase 2 + 3: peel persistence and image-export concerns out of
HeatmapSurveyViewModel into focused, Sendable value types.

- NetMonitor-iOS/Platform/HeatmapFileManager.swift (new, 104 LOC):
  owns the `.netmonsurvey` save/load dispatch (binary
  ProjectSaveLoadManager vs JSON fallback by extension), the
  autosave-URL convention against the Documents directory, and the
  static `listSavedProjects` / `deleteSavedProject` operations.
  `SavedSurveyInfo` moves here from the VM file as well.

- NetMonitor-iOS/Platform/HeatmapExporter.swift (new, 101 LOC):
  composites floor-plan + heatmap overlay + RSSI-coloured dots into
  a `UIImage` via `UIGraphicsImageRenderer`, plus the temp-file
  `.netmonsurvey` export path used by the share sheet. Carries its
  own static `rssiColor` mirror (the VM's instance method stays for
  the live canvas — same logic, two callers, no shared state).

- HeatmapSurveyViewModel: drops `private let projectManager:
  ProjectSaveLoadManager`, gains `private let fileManager:
  HeatmapFileManager` + `private let exporter: HeatmapExporter`,
  both injectable via the init with sensible defaults. The static
  `listSavedProjects` / `deleteSavedProject` are removed from the
  VM (and the now-redundant `SavedSurveyInfo` struct).
  saveProject / loadProject / autoSave / exportImage /
  exportProjectFile become thin delegators; the VM keeps the public
  API surface so view callers don't change.

- HeatmapProjectsView: switches the two static call sites to
  `HeatmapFileManager.listSavedProjects()` and
  `.deleteSavedProject(at:)`.

- HeatmapSurveyViewModelTests: updates the `listSavedProjects` test
  to hit `HeatmapFileManager` directly. `SavedSurveyInfo` is still
  in scope via the new file.

VM dropped from 655 → 543 LOC (overall #197 reduction: 714 → 543,
roughly -24%). NetMonitorCore `swift build` clean. iOS target
compile + Swift Testing run to be verified on the Mac mini node.

Phase 4 (RoomScanCoordinator) not pursued — calibration is
intrinsically coupled to surveyProject mutation; documented as
'won't-do' in the #197 comment.

Refs #197.
